+++ src/components/i18n/en.ts
@@ -23,32 +23,27 @@ export const en: i18n = {
         {
             id: 0,
             name: 'Warning Zones',
-            desc:
-                'In these Zones, which may not necessarily appear on the DJI GO map, users will be prompted with a warning message. Example Warning Zone: Class E airspace'
+            desc: 'In these Zones, which may not necessarily appear on the DJI GO map, users will be prompted with a warning message. Example Warning Zone: Class E airspace'
         },
         {
             id: 1,
             name: 'Authorization Zones',
-            desc:
-                'In these Zones, which appear blue in the DJI GO map, users will be prompted with a warning and flight is limited by default. Authorization Zones may be unlocked by authorized users using a DJI verified account.'
+            desc: 'In these Zones, which appear blue in the DJI GO map, users will be prompted with a warning and flight is limited by default. Authorization Zones may be unlocked by authorized users using a DJI verified account.'
         },
         {
             id: 2,
             name: 'Restricted Zones',
-            desc:
-                'In these Zones, which appear red the DJI GO app, users will be prompted with a warning and flight is prevented. If you believe you have the authorization to operate in a Restricted Zone, please contact flysafe@dji.com or Online Unlocking.'
+            desc: 'In these Zones, which appear red the DJI GO app, users will be prompted with a warning and flight is prevented. If you believe you have the authorization to operate in a Restricted Zone, please contact flysafe@dji.com or Online Unlocking.'
         },
         {
             id: 3,
             name: 'Enhanced Warning Zones',
-            desc:
-                'In these Zones, you will be prompted by GEO at the time of flight to unlock the zone using the same steps as in an Authorization Zone, but you do not require a verified account or an internet connection at the time of your flight.'
+            desc: 'In these Zones, you will be prompted by GEO at the time of flight to unlock the zone using the same steps as in an Authorization Zone, but you do not require a verified account or an internet connection at the time of your flight.'
         },
         {
             id: 4,
             name: 'Regulatory Restricted Zones',
-            desc:
-                'Due to local regulations and policies, flights are prohibited within the scope of some special areas. (Example：Prison)'
+            desc: 'Due to local regulations and policies, flights are prohibited within the scope of some special areas. (Example：Prison)'
         },
         {
             id: 5,
@@ -58,26 +53,22 @@ export const en: i18n = {
         {
             id: 6,
             name: 'Altitude Zones',
-            desc:
-                'Altitude zones will appear in gray on the map. Users receive warnings in DJI GO, or DJI GO 4 and flight altitude is limited.'
+            desc: 'Altitude zones will appear in gray on the map. Users receive warnings in DJI GO, or DJI GO 4 and flight altitude is limited.'
         },
         {
             id: 7,
             name: 'Recommended Zones',
-            desc:
-                'This area is shown in green on the map. It is recommended that you choose these areas for flight arrangements.'
+            desc: 'This area is shown in green on the map. It is recommended that you choose these areas for flight arrangements.'
         },
         {
             id: 8,
             name: 'Approved Zones for Light UAVs(China)',
-            desc:
-                'For Approved Zones, pilots of light UAVs flying at an altitude of 120 m or less are not required to obtain permission to fly. Pilots who are planning to fly medium-sized UAVs in Approved Zones at an altitude higher than 120 m, or in GEO Zones other than Approved Zones, must obtain permission via UTMISS before taking off'
+            desc: 'For Approved Zones, pilots of light UAVs flying at an altitude of 120 m or less are not required to obtain permission to fly. Pilots who are planning to fly medium-sized UAVs in Approved Zones at an altitude higher than 120 m, or in GEO Zones other than Approved Zones, must obtain permission via UTMISS before taking off'
         },
         {
             id: 9,
             name: 'Densely Populated Area',
-            desc:
-                'This area is shown in red on the map. Under normal circumstances, the population of this area is more concentrated, so please do not fly over this area. (Example: Commercial Block)'
+            desc: 'This area is shown in red on the map. Under normal circumstances, the population of this area is more concentrated, so please do not fly over this area. (Example: Commercial Block)'
         }
     ],
     types: [